+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.10 branch
 ### 2.10.0
+* `HtmlHelper` class has been improved. The `tip` option has been replaced
+    with `help`. Checkboxes are displayed according to the browser;
 * `DropdownHelper` class has been completely rewritten and now provides
     `menu()`, `start()` and `end()` methods;
 * added tests for `DropdownHelper` class.